problem with api response and comment loader
