+++ like.js
@@ -1,4 +1,6 @@
-export const initEventListeners = ({ comment, renderComments }) => {
+import { comment } from "./main.js";
+
+export const initEventListeners = ({ renderComments }) => {
     const likeButtonElement = document.querySelectorAll('.like-button');
     for (const item of likeButtonElement) {
         item.addEventListener('click', (event) => {
@@ -11,7 +13,7 @@ export const initEventListeners = ({ comment, renderComments }) => {
                 comment[index].isLiked = true;
                 comment[index].likes++;
             }
-            renderComments({ comment, initEventListeners });
+            renderComments({ initEventListeners });
         });
     };
 };
